Move Scaladoc test to correct package
